+++ src/webOS/coreservices/Main/main.jsx
@@ -3,7 +3,6 @@ import React, { StrictMode, useState, useEffect } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import System from '../System/System.jsx';
-import SystemUI from '../SystemUI/SystemUI.jsx';
 import MissionControl from '../MissionControl/MissionControl.jsx';
 
 const Main = () => {
